Day 18 - In which I regret never learning about parsers 😬
Part 1 was kinda okay for what it was, but as I wrote it I just KNEW it
wasn't going to help me with Part 2.

I tried to get @tsandall to teach me how to implement operator
precedence properly and then farted around for 2 hours trying to be
elegant but I think I stopped thinking things through about 30 min into
that so in the end I went with the part 2 solution you see here in
main.go which I think might be some of the worst code I've ever written.
BUT HEY IT WORKS....