Day 20 - Awful code 🙈 but so proud 😊
Okay this code is awful and no one should look at it ever BUT you should
have seen how happy I was when this finally worked. I was working on
this from day 20 until christmas eve and when it finally worked it
basically made my year. I had 33 (I kid you not, THIRTY THREE) different
versions of this sitting around. In the end it turned out I made a silly
mistake in my understanding of what rotatoin would do and it's kind of
amazing I got part 1 at all - once I figured that out (SO MANY HOURS
LATER) I was finally able to get it to work! Now that I know that, I
kinda hope I'll go back one day and remove like 90% of this code, but
who are we kidding, it'll be 2021 advent of code before we know it :D